python-service-green/requirements.txt
backoff==2.2.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
dnspython==2.6.0
Flask==3.0.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
idna==3.6
importlib-metadata==6.11.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
opentelemetry-api==1.23.0
opentelemetry-distro==0.44b0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-flask==0.44b0
opentelemetry-instrumentation-pymongo==0.44b0
opentelemetry-instrumentation-wsgi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
packaging==23.2
protobuf==4.25.3
pymongo==4.6.2
python-dotenv
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0